fix: [##62801] adjust ownership calculation
Previously, ownership for updates was determined by looking at the new
object state, i.e. the data in the GraphQL input. Now, you must be an
owner of both the existing and new data. For example moving an
engagement from one org unit to another requires ownership of both the
new and old org unit.