fix(mo2ldap): [#63154] handle temporally disjoint addresses
This commit fixes an issue where the LDAP integration would fail to
load an address from MO if another address of the same type existed at a
different disjoint time in history.

With this fix it will only fail if the two addresses overlap in time and
thus actually conflict. This is similar to how engagements are handled.