fix(ldap_event_generator): [#63334] support non-standard generalized time
This commit fixes an issue with the LDAP event-generator, which did not
correctly query Active Directory for changes. The issue was only related
to Active Directory and not to standard LDAP, as Active Directory does
*NOT* implement "Generalized-Time" in accordance with RFC 4517 Section 3.3.13:
https://datatracker.ietf.org/doc/html/rfc4517#section-3.3.13

Rather it implements its own non-compliant version which REQUIRED the
(by the standard) OPTIONAL fraction part, additionally it also REQUIRES it
to be set specifically to `".0"`.

See: https://learn.microsoft.com/en-us/windows/win32/adschema/s-string-generalized-time

No error is produced by Active Directory if the standard "Generalized-Time" is used
for searches, instead it fails silently with no search results.

The issue was introduced in version 11.3.3 of the integration with !876
(380b586f), which fixed another issue with the timestamp generation code,
specifically solving a problem where changes could be lost due to lacking
sub-second prceision.

This commit fixes the issue by generating timestamps in the Active
Directory specific format, as it is a valid subset of the RFC 4517
format, and thus is valid for both Standard LDAP and Active Directory.