fix: [#63334] paginate ldap-event-generator results
This commit fixes an issue with the LDAP event-generator in which
changes may be lost when fetching more than 1000 entries (server default)
in a single iteration.

This may occur if the event-generator is fetching entries from far in
the past until the current, for instance when starting the integration
for the first time.

Without this change a lot of changes may be lost as all entries above
1000 are simply pruned away.

The issue was introduced in version 1.0.0 of the integration and was
merged before very incremental change to the integration was released
with !101 (51e8391f).

This commit fixes the issue by doing a paginated search rather than
doing a simple search when fetching events. Most of the time there will
only be a single page, however paginated searches work in all instances,
while simple searches will lose results when the number of entries exceed
the server default.