fix: [#63443] ensure `dn2cpr` can handle missing CPR numbers
This commit fixes an issue where `dn2cpr` would raise an exception
instead of returning `None` upon encountering a missing CPR number.

The issue was introduced in version 23.2.0 of the integration with !1045
(13af4a8d), which intended to remove CPR validation from the integration
in favor of using OS2mo to check. The change removed an exception
suppression context manager, which unintentionally introduced the issue
as the context manager both suppressed exceptions related to invalid CPR
numbers as well as missing ones, not just invalid ones as expected.

This commit fixes the issue utilizing `more_itertools.only` instead of
`more_itertools.one` and by explicitly handling the `None`-case.