test(mo2ldap): [#63425] reproduce `None` mapping issue
This commit introduces a couple of test-cases verifying a production
issue related to mapping `None` values to LDAP.

The issue arises when a `None` value from MO is mapped to a field in
LDAP. Sometimes the value is successfully written using the stringified
`"None"` value, while other times it fails due to unicode encoding:
```
UnicodeError: Unable to convert type NoneType to unicode: None
```

The behavior seems to depend on the target field in LDAP, i.e. whether
the field supports unicode or not, however for both cases the current
behavior is undesirable, as it either entirely blocks the write or
alternatively writes a useless value.

In the future a commit will modify the current behavior such that values
in LDAP will be cleared whenever a `None` is templated. We believe this
is the expected behavior in these scenarios, and thus will align the
integration with 'Principle of least astonishment' in this situation.