fix(mo2ldap): [#63425] clear fields when templates return `None`
This commit fixes an issue with the OS2mo to LDAP synchronization in
which mapping `None` values from MO would either produce Unicode
encoding errors or write `"None"` to LDAP depending on the field in
question.

The issue was introduced in version 1.0.0 of the integration and was
merged before very incremental change to the integration was released
with !101 (51e8391f).

This commit fixes the issue by transforming the templated `None` values
to empty lists effectively clearing the fields whenever a `None` is
templated out.