test: [#63443] reproduce `dn2cpr` issue
This commit introduces a test-case verifying a production issue related
to extracting the CPR numbers from LDAP accounts.

The issue arises when no CPR number is found on the LDAP account.
The expected behavior in this situation would be that the `dn2cpr`
function returns `None`, however it currently produces an exception:
```
ValueError: too few items in iterable (expected 1)
```
As it tries to extract the CPR number from the list value returned from
LDAP.