test(mo2ldap): [#63444] reproduce broken it-user link issue
This commit introduces a unit-test reproducing a production issue where
attempting to create LDAP users would be blocked by the code attempting
to create linking it-users in MO.

The issue arises whenever the integration attempts to create LDAP users
while the `LDAP_IT_SYSTEM` environmental variable is set.

The exception raised is an `NoObjectsReturnedException`, which is caused
by trying to read the LDAP UUID of the object to be created before it is
actually created.