fix(ldap2mo): [#63444] ensure it-user links are created
This commit fixes an issue where creation of LDAP users was blocked by
it-user links not be created correctly due to an order of operations
issue, where the it-user link attempted to read the not-yet created LDAP
user.

The issue was introduced in version 23.5.0 of the integration with !1056
(e68939ac), which refactored the username generator to generate usernames
as opposed to it adding accounts to LDAP. The change unintentionally broke
the creation of it-user links, as the it-user link code was based on the
assumption that the username generator created the accounts in LDAP, not
just generating the usernames for them.

This commit fixes the issue by reordering the operations, such that the
it-user link is again created after the LDAP user has been created, thus
restoring the functionality of the integration as it was before the
issue was introduced.