refactor: [#63462] replace generate_username override with _name_fixer override
This commit replaces the `generate_username` with a `_name_fixer`
override within the `AlleroedUserNameGenerator` as the current
`generate_username` override is identical to the implementation in the
base class except from some additional cleaning of the employee name.

The `generate_username` already calls the `_name_fixer` method
internally and as such we can simply extend this method instead.