fix: [#63462] ensure generate_common_name works on Active Directory
This commit fixes an issue where `generate_common_name` did not work as
expected on non-standard LDAP implementations such as Active Directory.

The issue arises from the fact that Common Name (CN) is a list attribute
under standard LDAP, but this is not the case in Active Directory.

This commit fixes the issue by wrapping the returned value in an
`ensure_list` call such that both standard and non-standard LDAP
implementations can be handled the same.