feat(api): [#63483] introduce transitory SD API endpoint
This commit introduces a transitory endpoint to enable the SD
integration to migrate away from the old AD integration to this
integration instead.

The API endpoint is transitory as the good solution is for SD to create
users in MO with a random UUID instead of relying on creating them with
the entryUUID as UUID, however this requires OS2sync to be reconfigured,
which in itself relies on this integration to have been deployed
creating a circular dependency, that this endpoint is meant to break.