feat(api): [#63478] introduce transitory CPR UUID API endpoint
This commit introduces a transitory endpoint to enable the `cpr_uuid.py`
script to migrate away from the old AD integration to this integration
instead.

The API endpoint is transitory as the good solution is to eliminate the
use of the script and the file it produces entirely. This however
requires utilizing the new Rollekatalog integration, and perhaps
updating the OPUS integration as well, both of which requires this
integration to have been deployed, creating a circulat dependency, that
this endpoint is meant to break.