feat: [#63549] add E.164 phone format validator
This commit introduces a new validator scope for addresses, the E.164
phone format. The goal of the new validator is to ensure that
phonenumbers are saved in a standarized format in LoRa, such that
consumers of the data can always assume the format and process it as
such.

The implementation uses `python-phonenumbers` which is a Python port of
`google/libphonenumber` which handles international phonenumbers much
better than our current `PHONE` scope regex can.