feat(graphql, audit): [#62192] introduce actor translation
This commit introduces a new field `actor_object` to the `registrations`
and `auditlog` collections, such that the `actor` UUID can be resolved
to the corresponding object.

An example query resolving all the different actor types:
```graphql
query {
  registrations {
    objects {
      actor_object {
        uuid
        ... on SpecialActor {
          details
        }
        ... on PersonActor {
          person {
            cpr_number
          }
        }
        ... on IntegrationActor {
          name
        }
      }
    }
  }
}
```
With this query human made changes are shown via the `PersonActor`
returning the `cpr_number` of the user that made the change, while
changes made by integrations will return the `IntegrationActor` result
containing the name of the integration, while all special cases are
handled via the `SpecialActor` result, that contains a details text
explaining the situation.

The below result shows an example of an `actor_object` for each type:
```json
{
  "data": {
    "registrations": {
      "objects": [
        {
          "actor_object": {
            "uuid": "42c432e8-9c4a-11e6-9f62-873cf34a735f",
            "details": "The change was made before actors were registered"
          }
        },
        {
          "actor_object": {
            "uuid": "d1fec000-baad-c0de-0000-004449504558",
            "name": "dipex"
          }
        },
        {
          "actor_object": {
            "uuid": "0fb62199-cb9e-4083-ba45-2a63bfd142d7",
            "person": {
              "cpr_number": "1901593554"
            }
          }
        }
      ]
    }
  }
}
```