perf(graphql): [#63320] ensure bulking of address type scope
This commit makes a change to the implementation of `_get_handler_object`
to ensure that `address_type` scopes are properly bulked and loaded via
the dataloader, avoiding an N+1 performance issue.

In the case of the following query:
```graphql
query {
  addresses(limit: 500) {
    objects {
      current {
        name
      }
    }
  }
}
```
Prior to this change 1+500 LoRa calls would be run.
One query to fetch all the addresses, and one for each result, to fetch
the scope of the `address_type` scope to determine how to interpret the
address data, i.e. how to construct the name.

After this change this does 1+1 LoRa calls, the first one being the same,
and the second one fetching all `address_type` scopes with a single LoRa
call.