feat(mo2ldap): [#63692] implement create_user_trees
This commit implements `create_user_trees`, a mechanism for deciding
whether a user should be created in LDAP or not based on the
organization unit that their primary engagement in MO is attached to.

This feature was present in the old AD integration, and was not ported
to the new one and is yet another case of missing equivalence between
the two integrations. A lack of equivalence that this commit attempts to
resolve.