feat(moapi): [#63692] add get_ancestors method to MOAPI
This commit adds the `get_ancestors` method to the `MOAPI`, a method
which given an organization unit UUID finds all ancestors of that
organization unit and returns their UUIDs as a list.

The method is added in preparation of the following commit, which
utilizes it to implement `create_user_trees`.