feat(mo2ldap): [#63762] introduce get_employment_interval
This commit introduces `get_employment_interval` a new template function
available to users of the integration.

The function returns a tuple with the first employment date across all
engagements and the last employment date across all engagements.

The function is useful for specifying when an LDAP account should stop
existing, as this is usually when the last engagement stop, rather then
when the current engagement stops, although this depends on the specific
requirements for the customer.