test: [#63814] reproduce engagement override issue
This commit introduces a test to reproduce a production issue, where the
LDAP integration will gladly override an entire engagement history in MO
with its own data instead of merely updating the current engagement
history with a single field.

While this behavior is dangerous it is the expected behavior from the
way the code is structured.