feat(ldap2mo): [#63814] introduce UUID mapper
This commit introduces a new mapping behavior to the integration, which
can be toggled on using the `USE_UUID_MAPPING` feature flag.

The new mapping behavior shifts the responsibility of mapping from LDAP
objects to MO objects away from the complex hardcoded behaviors inside of
the LDAP integration to the jinja templates instead.

With the new mapping behavior mapping is done via the `uuid` field on
the object mappings, with the `uuid` being set to the object to be
created / terminate or to `None` if the object should be created.

To project against unwanted creates, one can use the `skip_if_none`
template wrapped around the `uuid` template, such that the object is
skipped instead of being created, whenever the template evaluates to
`false`.

The goal with this new mapping behavior is to eventually get rid of the
old mapping behavior entirely by exposing helper functions to the jinja
environment as needed to make the mapping.
This is already required for terminations to work as expected.

The new feature flag is set to `False` by default, and thus must be
enabled for each customer to get the new behavior.