feat(graphql): [#60458] move inherit flag to manager resolver
This commit moves the `inherit` flag from the manager field on
org-units to the `manager_resolver` in preparation for the introduction
of the `exclude_self` flag that is needed for both the frontend and the
new LDAP integration.

Moving the flag is really a refactoring in itself, changing nothing
about how the current flag functions, however the change is made as a
feature as it introduces the inherit flag on the manager top-level
collection as well.