feat(graphql): [#60458] introduce GraphQL version 23
This commit introduces GraphQL version 23 introduces a very minor
breaking change to the `managers` filter type when accessed from
within the `org_units` top-level collection.

The change is best explained with an example:
```graphql
query {
  org_units {
    objects {
      current {
        managers(
          filter: {}  # <-- The type of this filter is changed
        ) {
          uuid
        }
      }
    }
  }
}
```
Here the type of the filter is changed from `ManagerFilter` to
`OrgUnitsboundmanagerfilter`, while the type of the filter remains
unchanged (as `ManagerFilter`) in the following case:
```graphql
{
  managers(
    filter: {}  # <-- The type of this filter is NOT changed
	) {
    objects {
      current {
        uuid
      }
    }
  }
}
```

The change to `OrgUnitsboundmanagerfilter` serves to indicate that it is no
longer possible to pass `org_units` to the `managers` field in the first example.
While this was possible prior to the change, it would have no effect as the code
would simply override the given filter value with the contextual value.
Thus the only real breaking change is the type of the filter.

As such to migrate from GraphQL v22, simply update the type of the filter, and
if you were sending org_units in simply remove it, as it had no effect.