feat: [#61981] make `KLE.kle_number` a list (GraphQL v23)
GraphQL v23 changes the type of the `kle_number` field on `KLE` from
`Class!` to `[Class!]!`. It is now the same type as its sibling
`kle_aspects` field.

The previous typing incorrectly assumed that KLE-number classes were
static; valid from -infinity to infinity and never changing. Attempting
to read a `KLE` with a `kle_number` in the past or future would
therefore lead to an error, since the promise of always returning a
single KLE class could not be fulfilled. The field now returns a --
potentially empty -- list of classes valid in the given time period.