chore(graphql): [#60458] remove comment documenting alternative queries
This commit removes leftover queries documenting the performance
experiment done while introducing the organization unit child filter.

The filter was originally implemented thrice over using NOT IN, EXCEPT
and NOT EXISTS respectively.
The solution using NOT EXISTS has been kept for performance reasons, as
the NOT IN solution turned out to have horrible performance, while the
EXCEPT solution did not maintain the ordering of the result (and also
prohibited further WHERE-clauses).