feat(graphql): [#63829] add get_manager_person_uuid to jinja
This commit adds a new function `get_manager_person_uuid` to the jinja
template context. The function is used to get the current manager for
the provided `engagement_uuid` and can be used with `get_engagement_uuid`
or `load_mo_primary_engagement` to get the manager for any desired engagement.

The possibility to read managers is only half of the current customer
need, the other half being translating this manager reference into a DN
that can be written to the `manager` field in LDAP, however the
functionality offered by this template filter is likely to be useful for
other customers as these are onboarded and as such it is introduced as a
standalone filter in its own merge-request.

A future merge-request will include a function for translating any MO
UUID into their corresponding DN.