feat(graphql): [#63829] vacant manager filter
This commit adds support for finding vacant managers, by allowing users
to set `employee: null` on the manager filter.

This however is a breaking change, as setting `employee: null` on
GraphQL version 24 has the same effect as simply not including the
filter.
The following commit remidies this change by utilizing the version
introduced in the prior commit to add backwards compatibility.
The commits are split apart to ease the review process.

The feature is introduced by means of a hack in LoRa, introducing a new
magic value `urn:LORA-PLEASE-FIND-NULL-UUID-AND-URN`, which as the name
suggests instructs LoRa to find nows in which both the UUID and URN of
the relation are null, or finds missing rows.
This is exactly how vacant managers are encoded, and as such it serves
to implement the required filter.
For more information, check the massive comment explaining the behavior
in `search.py` within LoRa.

The magic value is similar to the previously introduced
`LORA-PLEASE-USE-IS-SIMILAR`, and should in the future be entirely
eliminiated in favor of an SQLAlchemy implementation of LoRa.

Co-Authored-By: Emil Madsen <emil@magenta.dk>