chore(mo2ldap): [#63829] remove writing references from _find_best_dn
This commit remove references to writing from `_find_best_dn` as the
`UserNameGenerator` now longer creates LDAP accounts when told to
generate a username candidate, but rather just creates a username
candidate.

The warnings and checks were simply leftovers from a time when the
`UserNameGenerator` did not have sane behavior.