test(graphql): [#63829] regression test for engagement.managers filter
This commit adds a test utilizing the `filter` on `engagement.managers`
to prove that the code implementing the filter is defective.

When the filter is provided the code fails with:
```
Org_UnitsBoundManagerFilter.__init__() got an unexpected keyword argument 'org_units'
```
As the filter type becomes an `Org_UnitsBoundManagerFilter` instead of
an ordinary `ManagerFilter` and as such, cannot have `org_units` set via
the `dataclasses.replace` method.