fix(graphql): [#63829] fix broken engagement.managers filter
This commit reverts the `engagement.managers` implementation to the state
it was in prior to the merge-request feedback, as the change introduced
during the review process turned out to be non-functional, while the
original state was fully functional.

An attempt was made to rectify the current state of the code by
converting the `Org_UnitBoundManagerFilter` to a normal `ManagerFilter`
within `engagement.managers`, such that the filter can would be seeded
correctly, however this approach would recursively dictionarify the
sub-filters, and thus break when those sub-filters would be resolved.