feat(graphql): [#63829] add get_person_dn to jinja
This commit adds a new function `get_person_dn` to the jinja template
context. The function is used to get the best known DN for the provided
MO `employee_uuid` and can be used with `get_manager_person_uuid` to get
the DN of a persons manager for writing the LDAP `manager` field.