feat(mo2ldap): [#63929] avoid spamming logs
This commit introduces a pre-write check to ensure that we do not
attempt to do "noop" "noop" in LDAP, as it was found that these writes
are not in fact treated as noop writes in Active Directory, but rather
as full writes that just happen to write the existing values, thus the
writes end up spamming the logs of the Active Directory server with
useless writes.

The check is implemented by reading the current state of the object we
intend to write, filtering the fields to write such that only fields
with changes are actually written.
If no fields are different, no write is performed.