test(mo2ldap): [#63930] ensure create-only attributes are possible
This commit adds a test to verify that we can make create-only
attributes using our Jinja templates. The test serves both to test that
it works, but also to document how to do it.

The template works by utilizing the `get_person_dn` template function to
fetch the current DN of the user, and if none exists adding new fields
to the template context, while if they exist doing nothing.
Thus it is possible to change behavior based on whether the current
operation is a create or update.