fix(username): [#63154] change LDAP login field
This commit changes the default LDAP login field from `distinguishedName`
to `uid` in correspondance with [`RFC2798`](https://www.rfc-editor.org/rfc/rfc2798)
and [`RFC1274`](https://www.rfc-editor.org/rfc/rfc1274).

Using `distinguishedName` never actually worked, as the generated usernames
are a single entry within the DN, not the entire DN and as such trying to
exclude full DNs from the username generator would never actually exclude
anything whatsoever.

However in the end this is also besides the point as the code was attempting
to read `distinguishedName` name as an attribute, when it in reality is not
an attribute at all, but rather the key of the object, thus the code was
not just non-standard, it was also broken in several ways.