chore(username): [#63154] remove AD assertion from AlleroedUserNameGenerator
This commit removes an assert from the AlleroedUserNameGenerator
allowing it to be used with Standard LDAP instead of being bound to
Microsoft Active Directory.

There was absolutely nothing in the code that was specific to Active
Directory, so there was no reason to for the assertion, and thus it has
been removed.