test(username): [#63154] add a bunch of tests for generate_username
This commit adds a bunch of tests for the `generate_username` jinja
template function, while also in effect testing the underlying
UsernameGenerator class, and the oddly named AlleroedUsernameGenerator.

The tests are added as the current tests for the UsernameGenerator are
mostly mocked unit-tests, while the introduced tests are all integration
tests that actually exercise the entire code-path.

This was discovered when making the AlleroedUsernameGenerator
configurable in `cdf745b7` / !1143, but unfortunately there was time at
the time to introduce tests, which is why they are added now after the
fact.