refactor(username): [#63930] move remove_vowels to UserNameGenerator
This commit moves the "remove_vowels" functionality in the `_name_fixer`
override of the `AlleroedUserNameGenerator` subclass to the base class
by hiding it behind a feature flag.

The goal of this change is to eventually eliminate the
`AlleroedUserNameGenerator` entirely, as it is used by multiple
customers, and as it currently gives those customers two behaviors, of
which they may only want one or the other.