refactor(username): [#63930] isolate code to fetch LDAP usernames
This commit moves the code to fetch LDAP usernames into a submethod
called `_get_existing_ldap_usernames` which is extracted from the
`_get_existing_usernames` method.

The change is made in preparation for the extraction of the code to
fetch MO usernames from the `AlleroedUserNameGenerator` such that this
code can be merged into the base-class `UserNameGenerator` in a method
called `_get_Existing_mo_usernames` both of which can be called from the
`_get_existing_usernames` method.