refactor(username): [#63930] remove _get_existing_usernames override
This commit removes the `_get_existing_usernames` method override from
the `AlleroedUserNameGenerator` by introducing a feature flag in the
base-class instead.

Thus at this point the `AlleroedUserNameGenerator` is nothing more than
a partial for the `__init__` method of the base-class `UserNameGenerator`
and can easily be eliminated in a future commit.