test(username): [#62009] prove existing_mo_usernames does not work in past and future
This commit adds a new parametrization axis to the
`test_generate_username_avoids_mo_taken_names` integration test, which
adds parametrization to the start and end dates of the it-users.

The `past` and `future` tests are added as `xfail`ed as they do not pass
with the state of the current code, proving the assumed bug.