fix(usernames): [#62009] ensure existing_mo_usernames works in past and future
This commit modifies the `read_all_ituser_user_keys_by_itsystem_uuid`
query such that it also searches for all past and future it-users,
instead of the current behavior which only returns the present ones.

The query is only used by the `_get_existing_mo_usernames` method of the
`UserNameGenerator` class, and now ensures that the comment within the
method is actually correct, i.e. that we actually check historic and
future username allocations too, instead of only present ones.

The commit also removes the `xfail`s that were introduced in the prior
commit to prove that the bug has now been resolved and that past and
future it-users now affect current username generation.