test(ldap_event_generator): [#63926] eliminate update_timestamp yield dependency
This commit refactors the `test_update_timestamp` test by eliminating
the usage of the yielded value of the `update_timestamp` async context
manager in favor of simply checking the database directly.

The refactoring is made in preparation of the inlining of the
`update_timestamp` async context manager into the `_generate_events`
function, which does not yield, nor return any values.