test(ldap_event_generator): [#63926] loop fold test_update_timestamp test
This commit loop folds the `test_update_timestamp` to avoid duplicate
code in test and to ensure that the test holds for cases larger than
two.

In addition it is done in preparation for another refactoring, which
removes the use of the `update_timestamp` async context manager yielded
value within the test.