test(ldap_event_generator): [#63926] eliminate update_timestamp async with dependency
This commit refactors the `test_update_timestamp` test by eliminating
the usage of the `async with` block from the `update_timestamp` async
context manager. To do so the expected test values must be changed
slightly as we now get the values prior to the full update, not the
values read before the update itself is done.

The refactoring is made in preparation of the inlining of the
`update_timestamp` async context manager into the `_generate_events`
function, which is a function and thus not an async context manager.