fix(ldap_event_generator): [#63926] use modifyTimestamp for last run
This commit modifies the time source for the last run database from
being the integrations `now` time, to instead being the time modify time
of the newest change event returned from the LDAP server.

This eliminates the risk of issues if the clock of the integration server
are out of sync with the clock of the LDAP server, for instance if NTP
issues occur, as we now compare LDAP server times with other LDAP server
times, instead of comparing LDAP server times with integration times.

If no events occur we simply reuse the old last run time till one occurs,
gradually increasing our search window over time.