feat(config)!: [#63928] disallow multiple domain controllers
This commit adds a validator to the config to ensure that at most one
domain controller is configured, as running against multiple is complex
and not currently handled gracefully by the integration.

In the future we may add proper support for multiple domain controllers,
but for now we simply disallow it until we find that it is needed.

BREAKING CHANGE:

The integration no longer supports multiple domain controllers.
To migrate to this version, simply remove all but one domain controller
for the configuration.