chore(api): [#64003] add horrible and dangerous cleanup endpoint
This commit adds a horrible and dangerous cleanup endpoint with the URL
`/danger/purge_addresses` which will remove all but one given address
and trigger a resynchronization of the affected user.

The endpoint should never have been needed, but due to bugs / poor design
in prior versions of the LDAP integration customer data may be in a bad
step that requires aggressive cleanup.