chore(username): [#63930] introduce username generator configuration variables
This commit introduces two new configuration variables for the
`UserNameGenerator`, namely `remove_vowels` and `disallow_mo_usernames`.

If `remove_vowels` is set it will remove vowels from surnames during
the username calculation.

If `disallow_mo_usernames` is set it will check OS2mo for usernames that
already have been used in the past or that are assigned in the future,
such that these usernames cannot be reused.
The `existing_usernames_itsystem` configuration variable designates the
it-system whereunder the users are checked for this purpose.