fix!: [#64040] stop creating addresses with values as user_keys
BREAKING CHANGE: The previous default behaviour for creating addresses
was to write the new value for the address as the user_key. This commit
stops that behaviour and allingns it with other objects in MO where the
user_key can be given as input to the request or it is set to the uuid
of the object.