fix(ldap_event_generator): [#63926] avoid spamming LDAP events
This commit fixes an issue where the LDAP event generator would spam the
same UUID endlessly onto the queue filling up the queue faster than it
could be processed until eventually the server would run out of desk and
crash spectacularly.

The issue was introduced in version 29.0.2 of the integration with !1161
(4e0f3e96), which fixed timing desync issues discovered when the
integration was deployed to a customer without NTP synchronization
between the LDAP server and the integration server.
This change modified the event generator, such that it would compare the
LDAP servers modifyTimestamp with itself, instead of relying on the
modifyTimestamp of the LDAP server to be aligned with the local time of
the integration.

The impact of the issue is likely minimal as the change has only been
deployed to a single customer, and as the queue likely only grows with a
single UUID per `poll_time` (typically 5 seconds).
The issue was noticed swiftly as it was manually deployed for testing.
Additionally the issue does not negatively affect the data integrity, as
it simply retriggers unnecessary resynchronizations over and over again,
which all turn out to be noops after the initial synchronization.

This commit fixes the issue by refusing to send out events iff the UUIDs
and modifyTimestamp of the current iteration of the event generator
matches the UUIDs and modifyTimestamp of the previous iteration.
As such events are only emitted if either the modifyTimestamp or the
UUIDs are changed between iterations.

The state to compare is stored in the database such that it survives
restarts of the integration, however this change required a change to
the `LastRun` database table, which unfortunately is not possible with
the FastRAMQPI framework, as it does not support Alembic, in lieu of
this support, the `LastRun` database table has had its table name
modified, such that the integration on start-up will issue a new
`CREATE TABLE` command for the new name, and as a consequence also
requeue *all* UUIDs in LDAP.
We do not consider this an issue, as there is no guarantee that the old
event-generator has correctly generated all the needed events, and as
such a full synchronization is in fact desirable.

The commit also includes a piece of code to remove the old database
table in order to limit confusion in future as to why two `LastRun`
tables exist. This code can be removed in future release.