fix(ldap_event_generator): [#63926] adhere to papal bull inter gravissimas anno 1582
This commit fixes an issue where the LDAP event generator would not
generate any events when running against Active Directory, if its
database had been initialized using Python's `datetime.min`.

The issue was introduced in version 29.1.3 of the integration with !1174
(3e5ab6e8), which fixed an event spamming incident discovered when the
integration was deployed to a customer following its redesign to deal
with NTP desynchronization in version 29.0.2 of the integration introduced
with !1161 (4e0f3e96).
The change modified the event generator to detect the situation that
would emit events endlessly becoming a noop when the situation occured.
Part of this change involved rotating the current database table for a
new one in lieu of migration support in the FastRAMQPI framework.
This also had the side-effect of reinitializing the table, such that
the existing timestamps were replaced with Python's `datetime.min`
aiming to trigger a full resynchronization of all data.

However no resynchronization was observed when deployed to test on the
customers Active Directory server, as the result set simply came back
empty.

Experimental lookups in Active Directory lead us to the fact that only
queries post the 1st of January 1601 produced any results, which with a
bit of online searching produced the explaination that this date was used
for the FILETIME start date within Windows NTs filesystem.

The date was chosen due to the papal bull inter gravissimas anno 1582,
aka the Gregorian reform to the Christian calendar, which went into effect
on the 15th of October 1582.

1602 was chosen, as it is the first year of the 400-year leap-year cycle
that was active when it was being designed.
The argument for this choice seems to be that choosing 1601 instead of 1582
(as RFC4122 does for UUIDs) simplifies the math for handling leap-years.

This means that the 1st of January 1601 is the date for which all Microsoft
Windows file times are calculated, and by extension the reference for all
Active Directory timestamps.

As such to support finding all entries in Active Directory when searching by
the modifyTimestamp, the initial search must take in a timestamp
corresponding to midnight on the 1st of January 1601.
Picking a time before this simply yields no results as Active Directory does
not handle dates before this, while picking a time after this date runs the
risk of missing elements whose modifyTimestamp is "NULL", as "NULL" is
commonly encoded as zero, and thus as midnight on the 1st of January 1601.

Thus to support Microsoft Active Directory, the database initialization
value cannot be Python's `datetime.min` but must instead be midnight of
the 1st of January 1601. - No other datetime will work correctly.

This commit changes the default value while rotating the database table
to ensure that the new default is used for all customers, while also
triggering a full resynchronization, this time hopefully succesfully so
even on Active Directory.

The commit also includes a piece of code to remove the old database
table in order to limit confusion in future as to why two `LastRun`
tables exist. The code to remove the prior "old old" database table
is kept in place for now, as that change happend only hours prior,
and thus have not had the chance to go by all deployments yet.
This code can be removed in future release.