chore(ci): [#64052] require 100% code coverage
This commit sets the required code-coverage to 100% such that all
newly added code must be tested (or atleast executed) by the test-suite.

The previous value of 80% was deemed undesirable (just as any other non
100% value), as it could block removal of dead (but tested) code, if
such a removal would bring the overall code coverage below 100%.

Additionally, it does not seem like a high bar to require that the code
we write is executed by the test suite. Having the requirement that the
code is executed does not imply that it is tested or even well-tested,
simply that it runs without crashing unexpectedly.

Unfortunately we cannot spare the time to introduce tests for all the code
that is currently not tested (nor executed), as the technical debt has
become insurmountable at this point, so rather than adding a bulkload of
tests up front, we will add `pragma: no cover` to ignore untested code
temporarily, to artificially reach 100% code coverage.

In the future all of these `pragma: no cover`s should be removed in
favor of testing of the code.